Second

+++ src/main/java/com/soma/calculator/Calculator.java
@@ -2,23 +2,23 @@ package com.soma.calculator;
 
 public class Calculator {
 
-	public static double add(double a, double b) {
+	public static int add(int a, int b) {
 		return a + b;
 	}
 
-	public static double subtract(double a, double b) {
+	public static int subtract(int a, int b) {
 		return a - b;
 	}
 
-	public static double multiply(double a, double b) {
+	public static int multiply(int a, int b) {
 		return a * b;
 	}
 
-	public static double divide(double a, double b) {
+	public static int divide(int a, int b) {
 		return a / b;
 	}
 
-	public static double mod(double a, double b) {
+	public static int mod(int a, int b) {
 		return a % b;
 	}
